feat: isolate ads player logic, fixes and cleanup
