+++ examples/nextjs-with-typescript/pages/MuxNewsPlayer.tsx
@@ -50,7 +50,23 @@ function MuxNewsPlayerPage() {
       {sdkReady && 
         <MuxNewsPlayer
           allowAdBlocker={true}
-          videoList={relatedVideos} 
+          videoList={relatedVideos}
+          onAdRequest={() => console.log('Ad Request')}
+          onAdResponse={() => console.log('Ad Response')}
+          onAdImpression={() => console.log('Ad Impression')}
+          onAdBreakStart={() => console.log('Ad Break Start')}
+          onAdPlay={() => console.log('Ad Play')}
+          onAdPlaying={() => console.log('Ad Playing')}
+          onAdPause={() => console.log('Ad Pause')}
+          onAdFirstQuartile={() => console.log('Ad First Quartile')}
+          onAdMidpoint={() => console.log('Ad Midpoint')}
+          onAdThirdQuartile={() => console.log('Ad Third Quartile')}
+          onAdClick={() => console.log('Ad Click')}
+          onAdSkip={() => console.log('Ad Skip')}
+          onAdClose={() => console.log('Ad Close')}
+          onAdEnded={() => console.log('Ad Ended')}
+          onAdBreakEnd={() => console.log('Ad Break End')}
+          onAdError={() => console.log('Ad Error')}
         />}
     </>
   );
