+++ examples/nextjs-with-typescript/pages/mux-video-ads-react.tsx
@@ -90,7 +90,7 @@ function MuxVideoPage() {
           playsInline={true}
           preferPlayback="mse"
           allowAdBlocker={true}
-          adTagUrl="https://pubads.g.doubleclick.net/gampad/ads?iu=/21775744923/external/vmap_ad_samples&sz=640x480&cust_params=sample_ar%3Dpostonly&ciu_szs=300x250&gdfp_req=1&ad_rule=1&output=vmap&unviewed_position_start=1&env=vp&impl=s&correlator="
+          adTagUrl="https://pubads.g.doubleclick.net/gampad/ads?iu=/21775744923/external/single_preroll_skippable&sz=640x480&ciu_szs=300x250%2C728x90&gdfp_req=1&output=vast&unviewed_position_start=1&env=vp&impl=s&correlator="
         />}
         
         <div>
