+++ packages/mux-player-react/package.json
@@ -20,11 +20,8 @@
       "themes/*": [
         "./dist/types-ts3.4/themes/*.d.ts"
       ],
-      "lazy": [
-        "./dist/types-ts3.4/lazy.d.ts"
-      ],
-      "news": [
-        "./dist/types-ts3.4/news.d.ts"
+      "*": [
+        "./dist/types-ts3.4/*.d.ts"
       ]
     },
     "*": {
@@ -34,11 +31,8 @@
       "themes/*": [
         "./dist/types/themes/*.d.ts"
       ],
-      "lazy": [
-        "./dist/types/lazy.d.ts"
-      ],
-      "news": [
-        "./dist/types/news.d.ts"
+      "*": [
+        "./dist/types/*.d.ts"
       ]
     }
   },
@@ -69,6 +63,13 @@
       "import": "./dist/themes/*.mjs",
       "require": "./dist/themes/*.cjs.js",
       "default": "./dist/themes/*.cjs.js"
+    },
+    "./*": {
+      "types@<4.3.5": "./dist/types-ts3.4/*.d.ts",
+      "types": "./dist/types/*.d.ts",
+      "import": "./dist/*.mjs",
+      "require": "./dist/*.cjs.js",
+      "default": "./dist/*.mjs"
     }
   },
   "types": "./dist/types/index.d.ts",
@@ -82,13 +83,15 @@
   "scripts": {
     "clean": "shx rm -rf dist/",
     "dev:types": "npm run build:types -- -w",
-    "dev:esm": "npm-run-all --parallel build:esm:** -- --watch=forever",
-    "dev:cjs": "npm-run-all --parallel build:cjs:** -- --watch=forever",
+    "dev:esm": "npm-run-all --parallel 'build:esm:** -- --watch=forever'",
+    "dev:cjs": "npm-run-all --parallel 'build:cjs:** -- --watch=forever'",
     "dev": "npm-run-all --parallel dev:types dev:esm dev:cjs",
     "build:cjs:index": "esbuild src/index.tsx src/themes/*.ts --target=es2019 --bundle --sourcemap --metafile=./dist/cjs.json --format=cjs --loader:.css=text --outdir=dist --out-extension:.js=.cjs.js --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
     "build:esm:index": "esbuild src/index.tsx src/themes/*.ts --target=es2019 --bundle --sourcemap --metafile=./dist/esm.json --format=esm --loader:.css=text --outdir=dist --out-extension:.js=.mjs --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
     "build:esm:news": "esbuild src/news/index.tsx --target=es2019 --bundle --sourcemap --metafile=./dist/news/esm.json --format=esm --loader:.css=text --outdir=dist/news --out-extension:.js=.mjs --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
     "build:cjs:news": "esbuild src/news/index.tsx --target=es2019 --bundle --sourcemap --metafile=./dist/news/cjs.json --format=cjs --loader:.css=text --outdir=dist/news --out-extension:.js=.cjs.js --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
+    "build:esm:ads": "esbuild src/ads.tsx --target=es2019 --bundle --sourcemap --metafile=./dist/esm.ads.json --format=esm --loader:.css=text --outdir=dist --out-extension:.js=.mjs --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
+    "build:cjs:ads": "esbuild src/ads.tsx --target=es2019 --bundle --sourcemap --metafile=./dist/cjs.ads.json --format=cjs --loader:.css=text --outdir=dist --out-extension:.js=.cjs.js --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
     "build:cjs:lazy": "echo 'esbuild cjs does not support code-splitting. See https://esbuild.github.io/api/#splitting for details'",
     "build:esm:lazy": "esbuild src/lazy.tsx --splitting --target=es2019 --bundle --sourcemap --metafile=./dist/esm.lazy.json --format=esm --loader:.css=text --outdir=dist --out-extension:.js=.mjs --external:react --external:@mux/* --external:prop-types --define:PLAYER_VERSION=\"'$npm_package_version'\"",
     "build:types": "tsc",
