+++ packages/mux-player-react/src/declaration.d.ts
